func-approx/requirements.txt
numpy
gym
matplotlib
tensorflow-gpu
pandas
ray[rllib]
requests
psutil
gputil
